pyproject.toml
[project]
name = "pear-weather"
version = "0.1.0"
description = "PEAR: Equal Area Weather Forecasting on the Sphere"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "boto3>=1.40.55",
    "cdsapi>=0.7.5",
    "dill>=0.3.9",
    "duckdb>=1.2.0",
    "einops>=0.8.1",
    "filelock>=3.17.0",
    "gitpython>=3.1.44",
    "healpix>=2024.2",
    "healpy>=1.18.1",
    "ipdb>=0.13.13",
    "ipython>=8.32.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "plotext==5.2.8",
    "pytest>=8.3.4",
    "python-lsp-ruff>=2.2.2",
    "ruff>=0.9.6",
    "timm>=1.0.14",
    "torch==2.5.1",
    "tqdm>=4.67.1",
    "xarray>=2025.1.2",
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
